+++ src/util/metadata.ts
@@ -3,9 +3,5 @@ import type { App } from 'obsidian';
 export function getResolvedLinks(
     app: App,
 ): Record<string, Record<string, number>> {
-    return (
-        app.metadataCache as unknown as {
-            resolvedLinks: Record<string, Record<string, number>>;
-        }
-    ).resolvedLinks;
+    return app.metadataCache.resolvedLinks;
 }